requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
asgiref==3.7.2
    # via
    #   django
    #   django-cors-headers
certifi==2023.11.17
    # via requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
coverage==7.4.0
    # via -r requirements.in
cryptography==41.0.7
    # via social-auth-core
defusedxml==0.8.0rc2
    # via
    #   python3-openid
    #   social-auth-core
django==5.0.1
    # via
    #   -r requirements.in
    #   django-admin-honeypot-updated-2021
    #   django-axes
    #   django-cors-headers
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   djoser
    #   social-auth-app-django
django-admin-honeypot-updated-2021==1.2.0
    # via -r requirements.in
django-axes==6.3.0
    # via -r requirements.in
django-cors-headers==4.3.1
    # via -r requirements.in
django-environ==0.11.2
    # via -r requirements.in
django-filter==23.5
    # via -r requirements.in
django-ipware==6.0.3
    # via django-admin-honeypot-updated-2021
django-templated-mail==1.1.1
    # via djoser
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.3.1
    # via
    #   -r requirements.in
    #   djoser
djoser==2.2.2
    # via -r requirements.in
gunicorn==21.2.0
    # via -r requirements.in
idna==3.6
    # via requests
markdown==3.5.2
    # via -r requirements.in
oauthlib==3.2.2
    # via
    #   requests-oauthlib
    #   social-auth-core
packaging==23.2
    # via gunicorn
pillow==10.2.0
    # via -r requirements.in
postgis==1.0.4
    # via -r requirements.in
psycopg2==2.9.9
    # via -r requirements.in
pycparser==2.21
    # via cffi
pyjwt==2.8.0
    # via
    #   djangorestframework-simplejwt
    #   social-auth-core
python-ipware==2.0.1
    # via django-ipware
python3-openid==3.2.0
    # via social-auth-core
pytz==2023.3.post1
    # via djangorestframework
requests==2.31.0
    # via
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==1.3.1
    # via social-auth-core
social-auth-app-django==5.4.0
    # via djoser
social-auth-core==4.5.1
    # via social-auth-app-django
sqlparse==0.4.4
    # via django
tzdata==2023.4
    # via django
urllib3==2.1.0
    # via requests
whitenoise==6.6.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
